# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 2026.5.16-1108mdt
  description: QuantumShield — Post-quantum encrypted chat API (ML-KEM-1024 + ML-DSA-87)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /identity-codes:
    get:
      operationId: getIdentityCodes
      tags: [identity-codes]
      summary: List identity and invite codes owned by the current account
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Owned codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IdentityCode"

    post:
      operationId: postIdentityCodes
      tags: [identity-codes]
      summary: Claim a globally unique alias or invite code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIdentityCodeRequest"
      responses:
        "201":
          description: Code claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityCode"
        "409":
          description: Code already claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /identity-codes/search:
    get:
      operationId: getIdentityCodesSearch
      tags: [identity-codes]
      summary: Search active public identity codes
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching account codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /identity-codes/{codeId}:
    patch:
      operationId: patchIdentityCodesCodeId
      tags: [identity-codes]
      summary: Activate, deactivate, roll, or retime an owned code
      security:
        - bearerAuth: []
      parameters:
        - name: codeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIdentityCodeRequest"
      responses:
        "200":
          description: Code updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityCode"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /keys/upload:
    post:
      operationId: postKeysUpload
      tags: [keys]
      summary: Upload pre-key bundle (ML-KEM-1024 + ML-DSA-87)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreKeyBundle"
      responses:
        "200":
          description: Keys uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /keys/{userId}:
    get:
      operationId: getKeysUserId
      tags: [keys]
      summary: Fetch a user's pre-key bundle for key exchange
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pre-key bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreKeyBundle"
        "404":
          description: No keys found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    User:
      type: object
      required: [id, username, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
          description: Active identity code for compatibility with older clients; may be "sealed".
        primaryCode:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        avatarColor:
          type: string
          nullable: true
        kemPublicKey:
          type: string
          nullable: true
          description: Base64-encoded ML-KEM-1024 public key
        dsaPublicKey:
          type: string
          nullable: true
          description: Base64-encoded ML-DSA-87 public key
        createdAt:
          type: string
          format: date-time
...
    IdentityCode:
      type: object
      required: [id, code, kind, visibilityScope, active, useCount, createdAt, updatedAt]
      properties:
        id:
          type: string
        code:
          type: string
        kind:
          type: string
          enum: [alias, invite]
        visibilityScope:
          type: string
          enum: [public, invited_by_you, invited_you, mutuals, disabled]
        active:
          type: boolean
        maxUses:
          type: integer
          nullable: true
        useCount:
          type: integer
        allowedCodes:
          type: array
          nullable: true
          items:
            type: string
        expiresAt:
          type: string
          format: date-time
          nullable: true
        rolledAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateIdentityCodeRequest:
      type: object
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 32
          nullable: true
        kind:
          type: string
          enum: [alias, invite]
          default: alias
        visibilityScope:
          type: string
          enum: [public, invited_by_you, invited_you, mutuals, disabled]
          default: public
        ttlSeconds:
          type: integer
          nullable: true
          description: Optional lifetime such as 300, 3600, 86400, 2592000, 31536000, or 315360000.
        maxUses:
          type: integer
          nullable: true
        allowedCodes:
          type: array
          nullable: true
          items:
            type: string
    UpdateIdentityCodeRequest:
      type: object
      properties:
        active:
          type: boolean
          nullable: true
        visibilityScope:
          type: string
          enum: [public, invited_by_you, invited_you, mutuals, disabled]
          nullable: true
        ttlSeconds:
          type: integer
          nullable: true
        maxUses:
          type: integer
          nullable: true
        allowedCodes:
          type: array
          nullable: true
          items:
            type: string
        confirmLastHandleDisable:
          type: boolean
          nullable: true
          description: Required only after triple confirmation and fresh re-authentication when disabling the last active handle.
...
    PreKeyBundle:
      type: object
      required: [kemPublicKey, dsaPublicKey, kemSignature]
      properties:
        kemPublicKey:
          type: string
          description: Base64-encoded ML-KEM-1024 public key
        dsaPublicKey:
          type: string
          description: Base64-encoded ML-DSA-87 public key
        kemSignature:
          type: string
          description: ML-DSA-87 signature over the KEM public key
        oneTimePreKeys:
          type: array
          nullable: true
          items:
            type: object
            required: [keyId, publicKey, signature]
            properties:
              keyId:
                type: string
              publicKey:
                type: string
              signature:
                type: string
...